Move keys function to main module (#27)
