+++ dynamo_pandas/__init__.py
@@ -1,4 +1,5 @@
 from .dynamo_pandas import get_df
+from .dynamo_pandas import keys
 from .dynamo_pandas import put_df
 from .dynamo_pandas import to_df
 from .dynamo_pandas import to_item
@@ -6,4 +7,4 @@ from .dynamo_pandas import to_items
 
 __version__ = "0.1.0"
 
-__all__ = ["get_df", "put_df", "to_df", "to_item", "to_items", "__version__"]
+__all__ = ["get_df", "keys", "put_df", "to_df", "to_item", "to_items", "__version__"]
