+++ dynamo_pandas/transactions/__init__.py
@@ -1,8 +1,7 @@
 from .transactions import get_all_items
 from .transactions import get_item
 from .transactions import get_items
-from .transactions import keys
 from .transactions import put_item
 from .transactions import put_items
 
-__all__ = ["get_all_items", "get_item", "get_items", "keys", "put_item", "put_items"]
+__all__ = ["get_all_items", "get_item", "get_items", "put_item", "put_items"]
